requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
mysql-connector-python==8.2.0
Flask-Login==0.6.3
python-dotenv==1.0.0
PyJWT==2.8.0
sqlparse==0.4.4
